Дебаг

+++ hw1/src/main/java/inputformat/WordCountJob.java
@@ -45,7 +45,7 @@ public class WordCountJob extends Configured implements Tool {
         }
     }
 
-    public static class WordCounterReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
+    public static class WordCounterReducer extends Reducer<Text, LongWritable, Text, IntWritable> {
         @Override
         protected void reduce(Text word, Iterable<IntWritable> nums, Context context) throws IOException, InterruptedException {
             int sum = 0;
